python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toggl_github_sync"
version = "0.1.0"
description = "Sync Toggl time entries to GitHub worklog"
readme = "README.md"
authors = [
    {name = "User", email = "user@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.28.0",
    "PyGithub>=1.58.0",
    "pytz>=2022.1",
    "python-dotenv>=0.20.0",
    "apscheduler>=3.9.1",
    "sendgrid>=6.9.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=0.961",
    "flake8>=4.0.1",
    "pytest>=7.1.2",
    "black>=22.3.0",
    "isort>=5.10.1",
]

[tool.setuptools]
packages = ["toggl_github_sync"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
